// .stryker-tmp/sandbox-WtNoiv/src/types/shogun.ts
export enum CorePlugins {
  /** WebAuthn plugin */
  WebAuthn = "webauthn",
  /** Ethereum plugin */
  Web3 = "web3",
  /** Bitcoin wallet plugin */
  Nostr = "nostr",
  /** OAuth plugin */
  OAuth = "oauth",
}
// ...
export interface WebauthnConfig {
  /** Enable WebAuthn */
  enabled?: boolean;
  /** Relying party name */
  rpName?: string;
  /** Relying party ID */
  rpId?: string;
}

/**
 * Shogun SDK configuration
 */
export interface ShogunSDKConfig {
  gunInstance?: IGunInstance<any>;
  authToken?: string;
  scope?: string;
  peers?: string[];
  webauthn?: WebauthnConfig;
  web3?: {
    enabled?: boolean;
  };
  nostr?: {
    enabled?: boolean;
  };
  oauth?: {
    enabled?: boolean;
    usePKCE?: boolean;
    allowUnsafeClientSecret?: boolean;
    providers?: Record<string, any>;
  };
  timeouts?: {
    login?: number;
    signup?: number;
    operation?: number;
  };
  plugins?: {
    autoRegister?: ShogunPlugin[];
  };
}